Take into account 'not in position' when making guesses
This fixes a logical deficiency in the core algorithm. When you get
a yellow (in-word) response in wordle, this also tells you that
the letter can't be in that exact position in the target word.
This is now considered in the algorithm by tracking it in the game state
and using this to choose guesses.